+++ nym-vpn-app/src-tauri/src/fs/mod.rs
@@ -1,4 +1,5 @@
 pub mod app;
+pub mod app_discovery;
 pub mod config;
 pub mod path;
 pub mod util;