+++ nym-vpn-app/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Lewes protocol settings toggle
+- Split tunneling for Windows
+
 ## [1.27.0] - 2026-03-31
 
 ### Added